move hpke public key handler out of internal (#4065)
